Moved resources and implementations for System.Text.Json for reuse across projects.
