+++ src/Umbraco.Cms.Api.Management/DependencyInjection/MvcBuilderExtensions.cs
@@ -2,7 +2,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Umbraco.Cms.Api.Management.Configuration;
+using Umbraco.Cms.Api.Common.Configuration;
 
 namespace Umbraco.Cms.Api.Management.DependencyInjection;
