+++ src/Umbraco.Cms.Api.Management/ManagementApiComposer.cs
@@ -17,8 +17,8 @@ using Umbraco.Cms.Core.Configuration.Models;
 using Umbraco.Cms.Core.DependencyInjection;
 using Umbraco.Cms.Api.Common.Configuration;
 using Umbraco.Cms.Api.Management.DependencyInjection;
-using Umbraco.Cms.Api.Management.Json;
 using Umbraco.Cms.Api.Management.OpenApi;
+using Umbraco.Cms.Infrastructure.Serialization;
 using Umbraco.Cms.Web.Common.ApplicationBuilder;
 using Umbraco.Extensions;
 using Umbraco.New.Cms.Core;
